scripts

+++ srcs/requirements/wordpress/tools/script.sh
@@ -1,29 +1,22 @@
 #!/bin/bash
-service php8.2-fpm start
+set -e
 
 DB_PASS=$(cat /run/secrets/db_password)
 DB_ROOT_PASS=$(cat /run/secrets/db_root_password)
 
-DB_HOST=${WORDPRESS_DB_HOST}
-DB_USER=${WORDPRESS_DB_USER}
-DB_NAME=${WORDPRESS_DB_NAME}
-SITE_URL=${WORDPRESS_URL}
-SITE_TITLE=${WORDPRESS_TITLE}
-ADMIN_USER=${WORDPRESS_ADMIN_USER}
-ADMIN_EMAIL=${WORDPRESS_ADMIN_EMAIL}
+if [ ! -f /var/www/html/wp-config.php ]; then
+  wp config create --allow-root \
+    --dbname="$WORDPRESS_DB_NAME" \
+    --dbuser="$WORDPRESS_DB_USER" \
+    --dbpass="$DB_PASS" \
+    --dbhost="$WORDPRESS_DB_HOST"
 
-wp core download --allow-root --path=/var/www/html
-wp config create --allow-root \
---dbname=${DB_NAME} \
---dbuser=${DB_USER} \
---dbpass=${DB_PASS} \
---dbhost=${DB_HOST}
-
-wp core install --allow-root \
---url=${SITE_URL} \
---title="${SITE_TITLE}" \
---admin_user=${ADMIN_USER} \
---admin_password=${DB_ROOT_PASS} \
---admin_email=${ADMIN_EMAIL}
+  wp core install --allow-root \
+    --url="$WORDPRESS_URL" \
+    --title="$WORDPRESS_TITLE" \
+    --admin_user="$WORDPRESS_ADMIN_USER" \
+    --admin_password="$DB_ROOT_PASS" \
+    --admin_email="$WORDPRESS_ADMIN_EMAIL"
+fi
 
 exec php-fpm8.2 -F
